requirements.txt
#notebook update
#notebook==7.0.6
#これを入れないとnotebookでtqdmが表示エラーになる
ipywidgets==8.0.6
#その他
# wandb==0.12.9
wandb==0.16.4
python-box==5.4.1
faiss-gpu==1.7.2
grad-cam==1.3.1
ttach==0.0.3
lightning==2.1.0

jupyter
ipykernel
python-dotenv
sentencepiece
seqeval
evaluate
kaggle
